bot/requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
gunicorn==20.0.4
psycopg2-binary==2.9.9
asgiref==3.8.1
Babel==2.9.1
Django==5.0.4
gino==1.0.1
middleware==1.2.3
multidict==6.0.4
SQLAlchemy==1.3.24
sqlparse==0.4.4
xlsxwriter==3.2.0
